move gui stuff to main folder/packages
