+++ src/main/resources/SharkToTheLeft.gif
