+++ src/main/resources/SharkToTheRight.gif
