+++ src/main/resources/sprites.jpg
